spa-cli/pyproject.toml
[tool.poetry]
name = "spa-cli"
version = "1.0.11"
description = "Un cli para manejar proyectos serverless en aws con python"
authors = ["David Cuy <david.cuy.sanchez@gmail.com>"]
readme = "README.md"
packages = [{include = "spa_cli"}]
homepage = "https://github.com/DavidCuy/spa-cli"
repository = "https://github.com/DavidCuy/spa-cli"
documentation = "https://github.com/DavidCuy/spa-cli"


[tool.poetry.scripts]
spa-cli = "spa_cli.cli:app"
spa = "spa_cli.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.1"
toml = "^0.10.2"
cookiecutter = "^2.6.0"
setuptools = "^80.9.0"
pydantic = "^2.11.9"
typer = "^0.19.1"
fastapi = {extras = ["standard"], version = "^0.116.1"}
tqdm = "^4.67.1"
mangum = "^0.19.0"
uvicorn = "^0.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
